--- src/hocs/withMap.js
@@ -1,81 +1,0 @@
-import React, {PureComponent, createRef} from 'react';
-import leaflet from 'leaflet';
-import {cardPropTypes} from '../utils/utils';
-import PropTypes from 'prop-types';
-
-const withMap = (Component) => {
-
-  class WithMap extends PureComponent {
-    constructor(props) {
-      super(props);
-      this._mapDiv = createRef();
-    }
-
-    componentDidMount() {
-      const city = [52.38333, 4.9];
-      const icon = leaflet.icon({
-        iconUrl: `/img/pin.svg`,
-        iconSize: [30, 30]
-      });
-
-      const zoom = 12;
-      this.map = leaflet.map(this._mapDiv.current, {
-        center: city,
-        zoom,
-        zoomControl: false,
-        marker: true
-      });
-      this.map.setView(city, zoom);
-
-      leaflet
-        .tileLayer(`https://{s}.basemaps.cartocdn.com/rastertiles/voyager/{z}/{x}/{y}{r}.png`, {
-          attribution: `&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors &copy; <a href="https://carto.com/attributions">CARTO</a>`
-        })
-        .addTo(this.map);
-      this.layerGroup = leaflet.layerGroup().addTo(this.map);
-      this.props.cards.forEach((card) =>
-        leaflet.marker(card.addressCoords, {icon}).addTo(this.layerGroup)
-      );
-    }
-
-    componentDidUpdate() {
-      const {hoveredId} = this.props;
-      const hoveredIcon = leaflet.icon({
-        iconUrl: `/img/pin-active.svg`,
-        iconSize: [30, 30]
-      });
-      let icon = leaflet.icon({
-        iconUrl: `/img/pin.svg`,
-        iconSize: [30, 30]
-      });
-      this.layerGroup.clearLayers();
-      this.markers = this.props.cards.forEach((card) => {
-        let newIcon = icon;
-        if (card.id === hoveredId) {
-          newIcon = hoveredIcon;
-        }
-        return leaflet.marker(card.addressCoords, {icon: newIcon}).addTo(this.layerGroup);
-      });
-    }
-
-    render() {
-      return (
-        <Component
-          reference={this._mapDiv}
-          nearPlace={this.props.nearPlace}
-        />
-      );
-    }
-  }
-
-  WithMap.propTypes = {
-    cards: PropTypes.arrayOf(cardPropTypes).isRequired,
-    nearPlace: PropTypes.bool,
-    hoveredId: PropTypes.number.isRequired
-  };
-
-  return WithMap;
-};
-
-export default withMap;
-
